+++ ydb/public/lib/ydb_cli/commands/ya.make
@@ -47,24 +47,23 @@ PEERDIR(
     ydb/library/formats/arrow/csv/table
     ydb/library/workload
     ydb/library/yaml_config/public
-    yql/essentials/public/decimal
-    ydb/public/sdk/cpp/src/library/operation_id
     ydb/public/lib/stat_visualization
-    ydb/public/lib/ydb_cli/common
     ydb/public/lib/ydb_cli/commands/command_base
     ydb/public/lib/ydb_cli/commands/interactive
+    ydb/public/lib/ydb_cli/commands/sdk_core_access
     ydb/public/lib/ydb_cli/commands/topic_workload
     ydb/public/lib/ydb_cli/commands/transfer_workload
     ydb/public/lib/ydb_cli/commands/ydb_discovery
-    ydb/public/lib/ydb_cli/commands/sdk_core_access
+    ydb/public/lib/ydb_cli/common
+    ydb/public/lib/ydb_cli/common
     ydb/public/lib/ydb_cli/dump
     ydb/public/lib/ydb_cli/dump/files
     ydb/public/lib/ydb_cli/import
     ydb/public/lib/ydb_cli/topic
-    ydb/public/sdk/cpp/src/client/draft
     ydb/public/sdk/cpp/src/client/bsconfig
     ydb/public/sdk/cpp/src/client/coordination
     ydb/public/sdk/cpp/src/client/debug
+    ydb/public/sdk/cpp/src/client/draft
     ydb/public/sdk/cpp/src/client/export
     ydb/public/sdk/cpp/src/client/import
     ydb/public/sdk/cpp/src/client/monitoring
@@ -75,6 +74,8 @@ PEERDIR(
     ydb/public/sdk/cpp/src/client/table
     ydb/public/sdk/cpp/src/client/topic
     ydb/public/sdk/cpp/src/client/types/credentials/login
+    ydb/public/sdk/cpp/src/library/operation_id
+    yql/essentials/public/decimal
 )
 
 GENERATE_ENUM_SERIALIZATION(ydb_benchmark.h)
